+++ WindowsFormsLocomotive/WindowsFormsLocomotive/WindowsFormsLocomotive.csproj
@@ -47,6 +47,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BaseLocomotive.cs" />
     <Compile Include="Direction.cs" />
     <Compile Include="FormLoco.cs">
       <SubType>Form</SubType>
@@ -54,9 +55,11 @@
     <Compile Include="FormLoco.Designer.cs">
       <DependentUpon>FormLoco.cs</DependentUpon>
     </Compile>
+    <Compile Include="ITransport.cs" />
     <Compile Include="Locomotive.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Vehicle.cs" />
     <EmbeddedResource Include="FormLoco.resx">
       <DependentUpon>FormLoco.cs</DependentUpon>
     </EmbeddedResource>
